Move delete_subtree out of CacheTree to bbound.cc
Move delete_subtree, which perhaps should be renamed, to bbound.cc
where we keep our algorithms, because variants with a permutation
map will need to modify entries to handle deleted nodes.

CacheTree destructor now uses destroy_subtree, which is like the
old delete_subtree (and oblivious to the permutation map).

Had to move permutation_insert, which uses delete_subtree,
to bbound.cc -- this could perhaps be refactored so that it goes
back to bbound.hh and the appropriate call to delete_subtree
happens by its caller (so permutation_insert would have to
also return a pointer to the permuted_node to pass to
delete_subtree, or NULL).

Could the permutation-related stuff (and queue-related) perhaps
be moved to separate files?  Haven't thought through what this
would look like.