Fixed linking error with global logger
Fixed vectorization of prefix lengths.

We should now be able to track the current size of the search space.
However, we are currently decrementing the number of prefix lengths when
we set a node to deleted, not when it is actually deleted, so the queue
size doesn't match up to the sum of the vector.