Fixed captured symmetry vector
Captured permutation map now acts directly on the VECTOR with a custom
hash function, leading to large memory and time savings over converting
it to a std::vector<bool>. It is now comparable to Prefix permutation
map in terms of memory and speed.

This only works with the GMP library currently.